refactor: Оптимизирована страница входа
